@@ -23,60 +23,63 @@
     "@clickhouse/client-web": "^0.3.1",
     "@headlessui/tailwindcss": "^0.2.0",
     "@hookform/resolvers": "^3.3.3",
-    "@radix-ui/react-accordion": "^1.1.2",
-    "@radix-ui/react-collapsible": "^1.0.3",
-    "@radix-ui/react-dialog": "^1.0.5",
-    "@radix-ui/react-dropdown-menu": "^2.0.6",
-    "@radix-ui/react-hover-card": "^1.1.1",
+    "@radix-ui/react-accordion": "^1.2.1",
+    "@radix-ui/react-collapsible": "^1.1.1",
+    "@radix-ui/react-dialog": "^1.1.2",
+    "@radix-ui/react-dropdown-menu": "^2.1.2",
+    "@radix-ui/react-hover-card": "^1.1.2",
     "@radix-ui/react-icons": "^1.3.0",
-    "@radix-ui/react-label": "^2.0.2",
-    "@radix-ui/react-navigation-menu": "^1.1.4",
-    "@radix-ui/react-popover": "^1.0.7",
-    "@radix-ui/react-select": "^2.0.0",
-    "@radix-ui/react-separator": "^1.0.3",
-    "@radix-ui/react-slot": "^1.0.2",
-    "@radix-ui/react-tabs": "^1.0.4",
-    "@radix-ui/react-toast": "^1.2.1",
-    "@radix-ui/react-tooltip": "^1.0.7",
-    "@tanstack/react-table": "^8.10.7",
-    "@tremor/react": "^3.15.0",
+    "@radix-ui/react-label": "^2.1.0",
+    "@radix-ui/react-navigation-menu": "^1.2.1",
+    "@radix-ui/react-popover": "^1.1.2",
+    "@radix-ui/react-select": "^2.1.2",
+    "@radix-ui/react-separator": "^1.1.0",
+    "@radix-ui/react-slot": "^1.1.0",
+    "@radix-ui/react-tabs": "^1.1.1",
+    "@radix-ui/react-toast": "^1.2.2",
+    "@radix-ui/react-tooltip": "^1.1.3",
+    "@tanstack/react-table": "^8.20.5",
+    "@tremor/react": "^3.18.3",
     "@uiw/react-heat-map": "^2.2.1",
     "@vercel/analytics": "^1.3.1",
     "@vercel/functions": "^1.4.0",
+    "babel-plugin-react-compiler": "^19.0.0-beta-6fc168f-20241025",
     "class-variance-authority": "^0.7.0",
-    "clsx": "^2.0.0",
+    "clsx": "^2.1.1",
     "cmdk": "^1.0.0",
     "dayjs": "^1.11.10",
     "dedent": "^1.5.3",
-    "lucide-react": "^0.453.0",
-    "next": "^14.2.3",
-    "react": "^18",
-    "react-dom": "^18",
-    "react-error-boundary": "^4.0.11",
-    "react-hook-form": "^7.49.2",
-    "react-resizable-panels": "^2.1.2",
-    "recharts": "^2.12.7",
+    "lucide-react": "^0.454.0",
+    "next": "^15.0.2",
+    "radix-ui": "^1.0.1",
+    "react": "^19.0.0-rc-0bc30748-20241028",
+    "react-dom": "^19.0.0-rc-0bc30748-20241028",
+    "react-error-boundary": "^4.1.2",
+    "react-hook-form": "^7.53.1",
+    "react-resizable-panels": "^2.1.6",
+    "recharts": "^2.13.1",
     "server-only-context": "^0.1.0",
-    "tailwind-merge": "^2.0.0",
+    "tailwind-merge": "^2.5.4",
+    "tailwindcss": "^3.4.14",
     "tailwindcss-animate": "^1.0.7",
-    "usehooks-ts": "^3.0.0",
-    "zod": "^3.22.4"
+    "usehooks-ts": "^3.1.0",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
     "@codecov/webpack-plugin": "^1.0.0",
     "@cypress/code-coverage": "^3.12.46",
     "@ianvs/prettier-plugin-sort-imports": "^4.2.1",
-    "@next/eslint-plugin-next": "^14.2.3",
-    "@types/jest": "^29.5.12",
+    "@next/eslint-plugin-next": "^15.0.0",
+    "@types/jest": "^29.5.14",
     "@types/node": "^20.11.28",
-    "@types/react": "^18.2.66",
-    "@types/react-dom": "^18",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
     "@typescript-eslint/parser": "^8.0.0",
     "autoprefixer": "^10.0.1",
-    "cypress": "13.15.1",
+    "cypress": "^13.15.1",
     "cypress-real-events": "^1.13.0",
     "eslint": "^8",
-    "eslint-config-next": "^14.2.3",
+    "eslint-config-next": "^15.0.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-cypress": "^4.0.0",
     "eslint-plugin-react": "^7.34.1",
@@ -89,9 +92,19 @@
     "prettier-plugin-organize-imports": "^4.0.0",
     "prettier-plugin-tailwindcss": "^0.6.0",
     "start-server-and-test": "^2.0.3",
-    "tailwindcss": "^3.4.3",
-    "ts-jest": "^29.1.4",
+    "ts-jest": "^29.2.5",
     "turbo": "^2.0.6",
     "typescript": "^5"
+  },
+  "overrides": {
+    "react": "^19.0.0-rc-0bc30748-20241028",
+    "react-dom": "^19.0.0-rc-0bc30748-20241028",
+    "react-is": "^19.0.0-rc-0bc30748-20241028",
+    "@tremor/react": {
+      "recharts": "^2.13.1"
+    },
+    "recharts": {
+      "react-is": "^19.0.0-rc-0bc30748-20241028"
+    }
   }
 }
