@@ -5,18 +5,20 @@ import {
 } from './breadcrumb'
 
 interface ClusterListProps {
-  params: {
+  params: Promise<{
     cluster: string
-  }
+  }>
   children: React.ReactNode
 }
 
 export const revalidate = 600
 
 export default async function ClusterTabListLayout({
-  params: { cluster },
+  params,
   children,
 }: ClusterListProps) {
+  const { cluster } = await params
+
   return (
     <div className="flex flex-col gap-5">
       <Suspense fallback={<ClusterListBreadcrumbSkeleton cluster={cluster} />}>
