@@ -1,12 +1,14 @@
 import { redirect } from 'next/navigation'
 
 interface PageProps {
-  params: {
+  params: Promise<{
     cluster: string
-  }
+  }>
 }
 
 // Redirects to the replicas status page.
-export default async function ClusterPage({ params: { cluster } }: PageProps) {
+export default async function ClusterPage({ params }: PageProps) {
+  const { cluster } = await params
+
   redirect(`/clusters/${cluster}/replicas-status`)
 }
