@@ -5,16 +5,16 @@ import { SampleData } from '../extras/sample-data'
 import { TableDDL } from '../extras/table-ddl'
 
 interface Props {
-  params: {
+  params: Promise<{
     host: number
     database: string
     table: string
-  }
+  }>
 }
 
-export default async function Dictionary({
-  params: { host, database, table },
-}: Props) {
+export default async function Dictionary({ params }: Props) {
+  const { host, database, table } = await params
+
   const engine = await engineType(database, table)
   if (engine !== 'Dictionary') return <></>
 
