@@ -3,11 +3,13 @@ import { redirect } from 'next/navigation'
 export const dynamic = 'force-dynamic'
 export const revalidate = 30
 
-export default function TablePage({
-  params: { host },
+export default async function TablePage({
+  params,
 }: {
-  params: { host: number }
+  params: Promise<{ host: number }>
 }) {
+  const { host } = await params
+
   // Redirect to the default database
   redirect(`/${host}/database/default`)
 }
