@@ -1,5 +1,3 @@
-import { unstable_noStore as noStore } from 'next/cache'
-
 import { ChartBackupSize } from '@/components/charts/backup-size'
 import { ChartCPUUsage } from '@/components/charts/cpu-usage'
 import { ChartDiskSize } from '@/components/charts/disk-size'
@@ -9,7 +7,7 @@ import { ChartMergeCount } from '@/components/charts/merge-count'
 import { ChartNewPartsCreated } from '@/components/charts/new-parts-created'
 import { ChartQueryCountByUser } from '@/components/charts/query-count-by-user'
 import { ChartTopTableSize } from '@/components/charts/top-table-size'
-import ChartKeeperException from '@/components/charts/zookeeper-exception'
+import { ChartKeeperException } from '@/components/charts/zookeeper-exception'
 import { ServerComponentLazy } from '@/components/server-component-lazy'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
 import { OverviewCharts } from './overview-charts'
@@ -17,10 +15,7 @@ import { OverviewCharts } from './overview-charts'
 export const dynamic = 'force-dynamic'
 export const revalidate = 5
 
-// Displays an overview of system metrics.
 export default async function Overview() {
-  noStore()
-
   return (
     <div>
       <OverviewCharts className="mb-10" />
