@@ -1,5 +1,5 @@
 import { redirectScoped } from '@/lib/scoped-link'
 
 export default async function TablePage() {
-  redirectScoped('/database/default')
+  await redirectScoped('/database/default')
 }
