@@ -2,10 +2,10 @@
 
 import { useEffect } from 'react'
 
-export function BackgroundJobs() {
+export function BackgroundJobs({ hostId }: { hostId: number }) {
   useEffect(() => {
     async function callCleanApi() {
-      await fetch('/api/clean')
+      await fetch(`/api/clean?hostId=${hostId}`)
     }
     callCleanApi()
   }, [])
